# lzbuffer: own output sink instead of holding ref