### Pull Request Overview

Changes `LZBuffer` trait to consume the output sink instead of holding a
reference to it. This makes it easier to store the sink and avoids
self-referential structs. It also makes sense for the buffer to own the
sink while it is performing decompression.

This also adds the methods `get_output` and `get_output_mut` to access
the output sink.

This pull request is needed by future versions of PR #51.

### Testing Strategy

This pull request was tested by...

- [ ] Added relevant unit tests.
- [ ] Added relevant end-to-end tests (such as `.lzma`, `.lzma2`, `.xz` files).

Should already be covered by existing tests.

### Supporting Documentation and References

Benchmarks:

**master**
```
test compress_65536                  ... bench:   3,187,997 ns/iter (+/- 863,604)
test compress_empty                  ... bench:       1,873 ns/iter (+/- 615)
test compress_hello                  ... bench:       2,072 ns/iter (+/- 608)
test decompress_after_compress_65536 ... bench:   3,624,633 ns/iter (+/- 1,528,450)
test decompress_after_compress_empty ... bench:       4,493 ns/iter (+/- 1,926)
test decompress_after_compress_hello ... bench:       5,452 ns/iter (+/- 5,584)
test decompress_big_file             ... bench:   8,809,873 ns/iter (+/- 5,368,703)
test decompress_huge_dict            ... bench:       4,430 ns/iter (+/- 2,042)
```

**owned_output_sink**
```
test compress_65536                  ... bench:   2,928,550 ns/iter (+/- 1,354,373)
test compress_empty                  ... bench:       1,513 ns/iter (+/- 512)
test compress_hello                  ... bench:       2,410 ns/iter (+/- 742)
test decompress_after_compress_65536 ... bench:   3,295,758 ns/iter (+/- 1,131,631)
test decompress_after_compress_empty ... bench:       3,893 ns/iter (+/- 1,449)
test decompress_after_compress_hello ... bench:       4,687 ns/iter (+/- 1,630)
test decompress_big_file             ... bench:   7,352,772 ns/iter (+/- 2,684,627)
test decompress_huge_dict            ... bench:       5,731 ns/iter (+/- 3,119)
```